requirements.txt
antlr4-python3-runtime==4.9.3
attributee==0.1.9
attrs==25.1.0
bidict==0.23.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==4.4.2
dominate==2.9.1
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
ftfy==6.3.1
hf-xet==1.1.2
huggingface-hub==0.32.2
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iopath==0.1.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazy-object-proxy==1.10.0
lazy_loader==0.4
llvmlite==0.44.0
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib==3.10.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.11.0.86
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
phx-class-registry==5.1.0
pillow==10.4.0
portalocker==3.1.1
proglog==0.1.10
psutil==7.0.0
py-cpuinfo==9.0.0
PyLaTeX==1.4.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
-e git+https://github.com/facebookresearch/sam2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4#egg=SAM_2
scikit-image==0.25.2
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.1
tifffile==2025.3.30
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.158
ultralytics-thop==2.0.14
urllib3==2.3.0
vot-toolkit==0.7.1
vot-trax==4.0.2
wcwidth==0.2.13
